moved aggregator claims class into claims file and removed seperate claims aggregator.
